+++ packages/js/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @pro6pp/infer-js
 
+## 0.0.2-beta.5
+
+### Patch Changes
+
+- Provide a fully-functional dropdown by default for infer-js and infer-react.
+- Updated dependencies
+  - @pro6pp/infer-core@0.0.2-beta.5
+
 ## 0.0.2-beta.4
 
 ### Patch Changes